pyproject.toml
# FILE: pyproject.toml
# VERSION: 1.0.0
# Build configuration for the TTS server project.
# Mirrors the requirement packs under profiles/packs/ as installable extras and
# exposes plugin discovery entry points for backends and model families.

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "tts-server"
version = "1.0.0"
description = "Local text-to-speech runtime with HTTP, Telegram, and CLI transports."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "tts-server contributors" },
]
keywords = [
    "tts",
    "text-to-speech",
    "qwen3",
    "piper",
    "omnivoice",
    "voice-cloning",
    "voice-design",
    "fastapi",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

# START_BLOCK_BASE_DEPENDENCIES
# Mirror profiles/packs/base/common.txt. Operators may continue to install via
# requirements*.txt; the pyproject extras give a parallel, PEP 621-aligned
# installation path that is also reachable from `pip install .[server]` etc.
dependencies = [
    "anyio==4.12.1",
    "audioread==3.1.0",
    "blobfile==3.1.0",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.1.8",
    "decorator==5.2.1",
    "filelock==3.19.1",
    "fsspec==2025.3.0",
    "h11==0.16.0",
    "hf-xet==1.4.2",
    "idna==3.11",
    "Jinja2==3.1.6",
    "joblib==1.5.3",
    "lazy_loader==0.4",
    "librosa==0.11.0",
    "llvmlite==0.46.0",
    "lxml==6.0.2",
    "MarkupSafe==3.0.3",
    "miniaudio==1.61",
    "msgpack==1.1.2",
    "numba==0.63.1",
    "numpy==2.3.5",
    "packaging==26.0",
    "platformdirs==4.5.1",
    "pooch==1.8.2",
    "protobuf==6.33.5",
    "pycparser==3.0",
    "pycryptodomex==3.23.0",
    "pyloudnorm==0.2.0",
    "pydantic==2.12.3",
    "PyYAML==6.0.3",
    "regex==2026.1.15",
    "requests==2.32.5",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.0",
    "sentencepiece==0.2.1",
    "shellingham==1.5.4",
    "sounddevice==0.5.3",
    "soundfile==0.13.1",
    "soxr==1.0.0",
    "standard-aifc==3.13.0",
    "standard-chunk==3.13.0",
    "standard-sunau==3.13.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "tqdm==4.67.1",
    "typer-slim==0.21.1",
    "typing_extensions==4.15.0",
    "urllib3==2.6.3",
    "accelerate==1.12.0",
]
# END_BLOCK_BASE_DEPENDENCIES

# START_BLOCK_OPTIONAL_DEPENDENCIES
[project.optional-dependencies]
# Module packs.
server = [
    "fastapi==0.121.1",
    "python-multipart==0.0.20",
    "uvicorn==0.38.0",
]
telegram = [
    "httpcore==1.0.9",
    "httpx==0.28.1",
]

# Family packs. Markers mirror profiles/packs/family/*.txt so the same
# resolution rules apply when installing through pip extras.
qwen = [
    "huggingface_hub==1.8.0; platform_system == 'Darwin'",
    "huggingface_hub==0.36.2; platform_system != 'Darwin'",
    "transformers==5.4.0; platform_system == 'Darwin'",
    "transformers==4.57.3; platform_system != 'Darwin'",
    "mlx==0.31.1; platform_system == 'Darwin'",
    "mlx-audio==0.4.1; platform_system == 'Darwin'",
    "mlx-lm==0.31.1; platform_system == 'Darwin'",
    "mlx-metal==0.31.1; platform_system == 'Darwin'",
    "qwen-tts==0.1.1; platform_system != 'Darwin'",
]
piper = [
    "onnxruntime==1.23.2",
    "piper-tts==1.4.2",
    "huggingface_hub==1.8.0",
    "torch>=2.6",
    "torchaudio>=2.6",
]
omnivoice = [
    "omnivoice==0.1.4",
]
qwen-fast = [
    "faster-qwen3-tts",
]

# Observability pack. Optional OpenTelemetry runtime - the bridge in
# core.services.telemetry stays no-op when these packages are absent.
telemetry = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]

# Tooling packs.
dev = [
    "pytest==8.3.0",
    "pytest-asyncio==0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=3.7.0",
    "types-PyYAML",
    "types-requests",
]
# END_BLOCK_OPTIONAL_DEPENDENCIES

[project.urls]
Repository = "https://github.com/maksglebov5757-bit/tts-server"
Issues = "https://github.com/maksglebov5757-bit/tts-server/issues"

# START_BLOCK_ENTRY_POINTS
# Plugin discovery surface. External packages can register additional backends
# or model families by exposing the same group keys in their own pyproject.
# The current built-in implementations are listed here so consumers can see the
# canonical extension shape; runtime auto-discovery is wired in a later phase.
[project.entry-points."tts_server.backends"]
torch = "core.backends.torch_backend:TorchBackend"
mlx = "core.backends.mlx_backend:MLXBackend"
onnx = "core.backends.onnx_backend:ONNXBackend"
qwen_fast = "core.backends.qwen_fast_backend:QwenFastBackend"

[project.entry-points."tts_server.model_families"]
qwen3_tts = "core.model_families.qwen3:Qwen3FamilyAdapter"
piper = "core.model_families.piper:PiperFamilyAdapter"
omnivoice = "core.model_families.omnivoice:OmniVoiceFamilyAdapter"
# END_BLOCK_ENTRY_POINTS

[project.scripts]
tts-cli = "cli.runtime:run_cli"
tts-telegram-bot = "telegram_bot.__main__:main"
tts-launcher = "launcher.main:main"

# START_BLOCK_HATCH_BUILD
[tool.hatch.build.targets.wheel]
packages = [
    "core",
    "server",
    "telegram_bot",
    "cli",
    "launcher",
    "profiles",
]

[tool.hatch.build.targets.wheel.shared-data]
# Profile pack files are read at runtime by the profile resolver.
"profiles/packs" = "profiles/packs"

[tool.hatch.build.targets.sdist]
include = [
    "/core",
    "/server",
    "/telegram_bot",
    "/cli",
    "/launcher",
    "/profiles",
    "/scripts",
    "/docs",
    "/tests",
    "/README.md",
    "/README.ru.md",
    "/AGENTS.md",
    "/requirements.txt",
    "/requirements-ci.txt",
    "/pytest.ini",
]
# END_BLOCK_HATCH_BUILD

# START_BLOCK_TOOL_PYTEST
[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: fast isolated tests for small logic units",
    "integration: adapter wiring and endpoint behavior tests without heavy inference",
    "smoke: opt-in real local runtime checks against a running server",
    "architecture: import boundary and layering checks",
    "anyio: async tests executed via anyio plugin",
]
# END_BLOCK_TOOL_PYTEST
